StoreCaseController
