+++ src/main/java/uk/gov/hmcts/reform/et/syaapi/controllers/StoreCaseController.java
@@ -11,10 +11,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.et.syaapi.annotation.ApiResponseGroup;
-import uk.gov.hmcts.reform.et.syaapi.models.SubmitStoredApplicationRequest;
+import uk.gov.hmcts.reform.et.syaapi.models.ClaimantApplicationRequest;
 import uk.gov.hmcts.reform.et.syaapi.models.UpdateStoredRespondToApplicationRequest;
 import uk.gov.hmcts.reform.et.syaapi.models.UpdateStoredRespondToTribunalRequest;
 import uk.gov.hmcts.reform.et.syaapi.service.StoredApplicationService;
+import uk.gov.hmcts.reform.et.syaapi.service.StoredRespondToApplicationSubmitService;
+import uk.gov.hmcts.reform.et.syaapi.service.StoredRespondToTribunalSubmitService;
 
 import javax.validation.constraints.NotNull;
 
@@ -28,25 +30,27 @@ import static uk.gov.hmcts.reform.et.syaapi.constants.EtSyaConstants.AUTHORIZATI
 public class StoreCaseController {
 
     private final StoredApplicationService storedApplicationService;
+    private final StoredRespondToApplicationSubmitService storedRespondToApplicationSubmitService;
+    private final StoredRespondToTribunalSubmitService storedRespondToTribunalSubmitService;
 
     /**
-     * Submits a stored Claimant Application.
+     * Store a Claimant Application.
      *
      * @param authorization jwt of the user
      * @param request       the request object which contains the claimant application passed from sya-frontend
      * @return the new updated case wrapped in a {@link CaseDetails}
      */
-    @PutMapping("/submit-stored-claimant-application")
-    @Operation(summary = "Submit a stored claimant application")
+    @PutMapping("/store-claimant-application")
+    @Operation(summary = "Store a claimant application")
     @ApiResponseGroup
-    public ResponseEntity<CaseDetails> submitStoredClaimantApplication(
+    public ResponseEntity<CaseDetails> storeClaimantApplication(
         @RequestHeader(AUTHORIZATION) String authorization,
-        @NotNull @RequestBody SubmitStoredApplicationRequest request
+        @NotNull @RequestBody ClaimantApplicationRequest request
     ) {
-        log.info("Received submit the stored claimant application request - caseTypeId: {} caseId: {}",
+        log.info("Received store claimant application request - caseTypeId: {} caseId: {}",
                  request.getCaseTypeId(), request.getCaseId()
         );
-        CaseDetails finalCaseDetails = storedApplicationService.submitStoredApplication(authorization, request);
+        CaseDetails finalCaseDetails = storedApplicationService.storeApplication(authorization, request);
         return ok(finalCaseDetails);
     }
 
@@ -67,7 +71,8 @@ public class StoreCaseController {
         log.info("Received submit respond to application request - caseTypeId: {} caseId: {}",
                  request.getCaseTypeId(), request.getCaseId()
         );
-        CaseDetails finalCaseDetails = storedApplicationService.submitRespondToApplication(authorization, request);
+        CaseDetails finalCaseDetails =
+            storedRespondToApplicationSubmitService.submitRespondToApplication(authorization, request);
         return ok(finalCaseDetails);
     }
 
@@ -88,7 +93,8 @@ public class StoreCaseController {
         log.info("Received submit respond to application request - caseTypeId: {} caseId: {}",
             request.getCaseTypeId(), request.getCaseId()
         );
-        CaseDetails finalCaseDetails = storedApplicationService.submitRespondToTribunal(authorization, request);
+        CaseDetails finalCaseDetails =
+            storedRespondToTribunalSubmitService.submitRespondToTribunal(authorization, request);
         return ok(finalCaseDetails);
     }
 }
