+++ src/main/java/uk/gov/hmcts/reform/et/syaapi/enums/CaseEvent.java
@@ -15,6 +15,7 @@ public enum CaseEvent {
     SUBMIT_CASE_DRAFT,
     UPDATE_CASE_SUBMITTED,
     SUBMIT_CLAIMANT_TSE,
+    STORE_CLAIMANT_TSE,
     SUBMIT_STORED_CLAIMANT_TSE,
     CLAIMANT_TSE_RESPOND,
     UPDATE_APPLICATION_STATE,
