--- src/main/java/uk/gov/hmcts/reform/et/syaapi/models/SubmitStoredApplicationRequest.java
@@ -1,18 +1,0 @@
-package uk.gov.hmcts.reform.et.syaapi.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Data;
-import lombok.extern.jackson.Jacksonized;
-
-@Data
-@Builder
-@Jacksonized
-public class SubmitStoredApplicationRequest {
-    @JsonProperty("case_id")
-    private String caseId;
-    @JsonProperty("case_type_id")
-    private String caseTypeId;
-    @JsonProperty("application_id")
-    private String applicationId;
-}
